+++ convex/users.ts
@@ -2,7 +2,6 @@ import { v } from 'convex/values';
 import { api, internal } from './_generated/api';
 import { Doc, Id } from './_generated/dataModel';
 import {
-  action,
   internalAction,
   internalMutation,
   internalQuery,
